requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
Authlib==1.3.2
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
flower==2.0.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
pytz==2024.2
redis==5.0.8
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.0.1
